+++ lib/core/router/app_router.dart
@@ -19,6 +19,7 @@ import '../../features/pengaturan/payment_methods_screen.dart';
 import '../../features/pengaturan/pengaturan_screen.dart';
 import '../../features/pengaturan/employee_screen.dart';
 import '../../features/pengaturan/expenses_screen.dart';
+import '../../features/pengaturan/tutup_kasir_screen.dart';
 import '../../features/pengaturan/printer_screen.dart';
 import '../../features/pengaturan/store_info_screen.dart';
 import '../../features/pengaturan/sync_screen.dart';
@@ -158,6 +159,9 @@ final routerProvider = Provider<GoRouter>((ref) {
               GoRoute(
                   path: 'pengeluaran',
                   builder: (_, __) => const ExpensesScreen()),
+              GoRoute(
+                  path: 'tutup-kasir',
+                  builder: (_, __) => const TutupKasirScreen()),
               GoRoute(
                   path: 'pegawai', builder: (_, __) => const EmployeeScreen()),
               GoRoute(